map search + note rework
added a search bar to the map select tab and also made hovering over markers display the legend alongside the note. ontop of that, the note or legend now sticks to the cursor instead of just being under the marker. (except in mobile where it just sticks to the side)